make odourFIeld compatible with lr by NOT making odourQuads but by making each Quad as large as world So, each reset event will use a complete map of a given quad
Reorganize positioning of objects. It is now in order ofbinary
00,01,10,11  but the ordering in setObjects is different(dunno why, made it conform with this) Have to fix it

wind and odour will be generated at init regardless of params. but will be used on frames based on params
reorganized the functions a bit to be able to handle the additional capacity of lr experiment calss in odour

rename odourField to odourQuadfield to be explicit
init self.pf to 0 in tunnels
addd todo for symetrical lighting